Update tests, fix bug
