+++ src/test/resources/retry/config-retry.yaml
@@ -1,0 +1,32 @@
+resilience4j:
+    retryConfigurations:
+      - name: constantRetry
+        maxAttempts: 1
+        intervalFunction:
+          type: constant
+          initialInterval: 10ms
+        ignoreExceptions:
+          - java.lang.RuntimeException
+        retryExceptions:
+          - java.lang.IllegalArgumentException
+      - name: randomizedRetry
+        maxAttempts: 2
+        intervalFunction:
+          type: randomized
+          initialInterval: 10ms
+          randomizationFactor: 0.5
+        ignoreExceptions:
+        retryExceptions:
+      - name: exponentialBackoffRetry
+        maxAttempts: 3
+        intervalFunction:
+          type: exponentialBackoff
+          initialInterval: 10ms
+          multiplier: 3.5
+      - name: exponentialRandomizedBackoffRetry
+        maxAttempts: 4
+        intervalFunction:
+          type: exponentialRandomBackoff
+          initialInterval: 10ms
+          multiplier: 2.5
+          randomizationFactor: 0.5
